Safari-609.1.11.1/LayoutTests/webgl/2.0.0/conformance2/reading/read-pixels-into-pixel-pack-buffer-expected.txt

CONSOLE MESSAGE: line 113: TypeError: Argument 7 ('pixels') to WebGL2RenderingContext.readPixels must be an instance of ArrayBufferView
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance2/reading/read-pixels-into-pixel-pack-buffer.html
[ 1: PASS ] context exists
[ 2: PASS ] getError was expected value: INVALID_OPERATION : should generate INVALID_OPERATION if pixel pack buffer is bound
[ 3: FAIL ] successfullyParsed should be true (of type boolean). Was undefined (of type undefined).
[ FAIL ] 1 failures reported
